Rename tests.
