+++ src/test/cbmc/Assert/UninitValid/fixme_main.rs
@@ -1,5 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+// rmc-verify-fail
+
 use std::mem::{self, MaybeUninit};
 
 fn main() {
