+++ src/test/cbmc/Assert/ZeroValid/fixme_main.rs
@@ -1,5 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+// rmc-verify-fail
+
 // The function zeroed() calls assert_zero_valid to mark that it is only defined to assign an
 // all-zero bit pattern to a type T if this is a valid value. So the following is undefined.
