+++ src/test/cbmc/BinOp_Offset/main_fail.rs
@@ -1,6 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
-
 // rmc-verify-fail
 
 include!("../../rmc-prelude.rs");
