+++ src/test/cbmc/DynTrait/nested_boxes_fail.rs
@@ -1,6 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
-
 // rmc-verify-fail
 
 // This test checks the size and align fields for 3-deep nested trait pointers. The
