+++ src/test/cbmc/Intrinsics/abort.rs
@@ -1,11 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
-
 // rmc-verify-fail
 
-use std::fs;
-use std::path::PathBuf;
-pub fn main() {
-    let buf = PathBuf::new();
-    let _x = fs::remove_dir_all(buf);
+#![feature(core_intrinsics)]
+use std::intrinsics;
+fn main() {
+    intrinsics::abort();
 }
