+++ src/test/cbmc/Pointers_Basic/fixme_from_raw.rs
@@ -1,5 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+// rmc-verify-fail
 
 fn main() {
     let address = 0x01234usize;
