+++ src/test/cbmc/Pointers_OutOfScopeFail/fixme_main.rs
@@ -1,5 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+// rmc-verify-fail
+
 fn main() {
     // declare pointer to integer
     let p_subscoped: *const u32;
