+++ src/test/cbmc/Serde/main.rs
@@ -1,5 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+// rmc-check-fail
+
 // We currently assert(false) for reify function pointer.
 // So this codegens but fails model checking.
