+++ src/test/cbmc/SizeAndAlignOfDst/main_fail.rs
@@ -5,7 +5,6 @@
 // Arc<Mutex<dyn Subscriber>>.
 // This test still fails with a final coercion error for
 // DummySubscriber to dyn Subscriber.
-
 // rmc-verify-fail
 
 use std::sync::Arc;
