+++ src/test/cbmc/Slice/slice_from_raw.rs
@@ -1,6 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
-
 // rmc-verify-fail
 
 use std::slice;
