+++ src/test/smack/basic/add.rs
