+++ src/test/smack/basic/arith_assume3.rs
