+++ src/test/smack/basic/mul.rs
