+++ src/test/smack/generics/generic_function1.rs
