+++ src/test/smack/generics/generic_function2.rs
