+++ src/test/smack/generics/generic_function3.rs
