+++ src/test/smack/generics/generic_function4.rs
