+++ src/test/smack/generics/generic_function5.rs
