+++ src/test/smack/overflow/mul_overflow.rs
