+++ src/test/smack/vector/vec11.rs
