+++ src/test/smack/vector/vec12.rs
